Placeholder for port to Zephyr running on ARM
The other ports in this project are all POSIX based.  e.g.,
the existing Zephyr port uses the 'POSIX' architecture.
This unfortunately doesn't work for an ARM based Zephyr
project like Puck; note that there is an option to enable
POSIX compatability in Zephyr for non-POSIX architectures,
but it's a subset, and IOTC will not compile against it.

Currently this port uses a dummy filesystem, and uses tls_socket
instead of linking a TLS library.  We'll need to move over
to using a memory filesystem; and TLS socket vs library is TBD.